refactor

+++ src/com/anitsuga/shop/api/meli/model/ValueStruct.java
@@ -2,10 +2,12 @@ package com.anitsuga.shop.api.meli.model;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ValueStruct {
 
-    private Integer number;
+    private BigDecimal number;
     private String unit;
 }
